Create product-overvieew folder and move product related components in there
