+++ client/src/components/App.jsx
@@ -3,7 +3,7 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import QandA from './Q-and-A/QandA.jsx';
 import RatingsReviews from './RatingsAndReviews/RatingsReviews.jsx';
-import Product from './Product.jsx';
+import Product from './Product-overview/Product.jsx';
 import { appContext } from '../contexts/index.js';
 
 export default function App() {
